ARTEMIS-5037: option to limit mirror propagation
Add a new option in the Mirror setting to prevent a broker from
propagating messages.

When working with a topology where 4 nodes form a square and each node
mirrors to its two neighbors, a message leaving a corner can reach the
opposite corner of the square by two different routes. This is causing
the ordering of message to get broken.

example:
1 <-> 2
^     ^
|     |
v     v
4 <-> 3

A message from a will reach 3 by 2 and 4. Message duplication checks
will prevent the message from being duplicated but won't help regarding
the order of the messages.

Using the new option to not forward messages coming from a link, we
break the possibilities to have two routes to reach the opposite corner.

On the example above, we ask 4 to not forward any messages coming from
1 if 1 is the primary sender of the messages.

This means that now, a message sent to 1 will reach 3 only by 2 and not
by 4.